[HAR-7889]: Sumo logic cv 24X7 log verification (#5245)
* work in progress for data collection for logs

* [HAR-7889]: end to end log ml analysis for 24X7

* [HAR-7889]: fix pr checks

* [HAR-7889]: fix integration tests

* Start heatmap logs

* Heat map API for logs

* Heatmap dashboard for logs

* Fix filtering

* end to end 24X7 log verification

* More things

* fix max log data collection minute

* more fixes

* fixes for validation

* Trigger new build

* optimize query to find hosts for a specifi cluster level

* Revert "More things"

This reverts commit f20346edf2e055d0f5504f483e3cbaab62d146db.

* Revert "Heatmap dashboard for logs"

This reverts commit 6db8e9bbb30de2f3684244b6c93e328781203743.

* Revert "Heat map API for logs"

This reverts commit 811a6229f844932587379623916df8e69cc2c2d6.

* Revert "Start heatmap logs"

This reverts commit 4cec4bca47e5516cd11c3c39876da3996735d2bd.

* [HAR-7889]: fixing tests

* [HAR-7889]: move data collection trigger to happen once every 90 seconds

* [HAR-7889]: do not clean up metric records yet

* Revert "Revert "Start heatmap logs""

This reverts commit c406f272865737e74dff7daa8a0d09634142ab36.

* Revert "Revert "Heat map API for logs""

This reverts commit 7b0addf9bc85b6d6d8427f28d557d8eddef3784f.

* Revert "Revert "Heatmap dashboard for logs""

This reverts commit 8670d0f9743138348462935b0f7df9251f6c0d37.

* Checkstyle

* [HAR-7889]: fix required indices

* Checkstyle and pmd

* fix build

* [HAR-7889]: review comments

+++ 71-rest/src/main/java/software/wings/service/impl/analysis/LogElement.java
@@ -19,5 +19,5 @@ public class LogElement {
   private long timeStamp;
   private int count;
   private String logMessage;
-  private int logCollectionMinute;
+  private long logCollectionMinute;
 }
