+++ 79-verification/src/main/java/io/harness/service/intfc/TimeSeriesAnalysisService.java
@@ -84,7 +84,7 @@ public interface TimeSeriesAnalysisService {
   void saveMetricTemplates(String appId, StateType stateType, String stateExecutionId,
       Map<String, TimeSeriesMetricDefinition> metricTemplates);
 
-  int getMaxCVCollectionMinute(String appId, String cvConfigId);
+  long getMaxCVCollectionMinute(String appId, String cvConfigId);
 
   long getLastCVAnalysisMinute(String appId, String cvConfigId);
